Have `cache_version` be the prefix in all cases, rather than the suffix.
This gives the `cache_version` the control it needs to effectively bust a
cache.  Without having the version _earlier_ in the cache key, stale caches
can still happen since CircleCI falls back to less and less specific
versions of the cache from right to left.

Also, this removes the possibility of it ever falling back to an
un-versioned cache which opens the door to using a permanently bricked cache
without control over bumping the version, or allows the caches to blur
between platforms (e.g., a Unix cache using a macOS cache).

This is also per the advice of the CircleCI documentation:

  https://circleci.com/docs/2.0/caching/#clearing-cache

Fixes: https://github.com/CircleCI-Public/rust-orb/issues/10